model/pyproject.toml
[tool.poetry]
name = "toxicity-analyzer-model"
version = "0.1.0"
description = "Toxicity analyzer model component"
authors = ["cbw"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
Flask = "^2.0.2"
Flask-Cors = "^3.0.10"
detoxify = "^0.4.0"
python-dotenv = "^0.19.2"
pytest = "^7.0.1"
futures3 = "^1.0.0"
numpy = "^1.22.2"
locust = "^2.8.2"
prometheus-client = "^0.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
